+++ src/LocaleRouteServiceProvider.php
@@ -2,6 +2,7 @@
 
 namespace CaribouFute\LocaleRoute;
 
+use CaribouFute\LocaleRoute\Locale\Route as LocaleRouteUrl;
 use CaribouFute\LocaleRoute\Routing\Router as LocaleRouter;
 use Illuminate\Support\ServiceProvider;
 
@@ -18,6 +19,7 @@ class LocaleRouteServiceProvider extends ServiceProvider
     public function register()
     {
         $this->app->bind('locale-route', LocaleRouter::class);
+        $this->app->bind('locale-route-url', LocaleRouteUrl::class);
         config('config/localeroute.php');
     }
 }