+++ changelog.md
@@ -7,10 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Coming soon]
 ### To add
+- Complete missing unit and functional tests for complete code coverage.
+
+## [1.0.0-alpha] - 2016-11-09
+### Added
 - *LocaleRoute::resource* method
 
-### To change
-- *other_locale* will keep current parameters by default
+### Changed
+- *locale_route*, *other_locale* and *other_route* keep current parameters by default when not declared.
 
 ## [1.0.0-alpha] - 2016-11-07
 ### Added